프로젝트 상세 페이지 빼고는 정리 완료
