+++ src/app/project/[projectId]/_component/Header/LeftSection.tsx
@@ -1,3 +1,5 @@
+import { SmallUserImgIcon } from "@/components/icons/icons";
+
 export default function LeftSection() {
     return (
         <div>
@@ -19,19 +21,7 @@ export default function LeftSection() {
 
             <div className="flex items-center gap-2">
                 <div className="flex gap-2 items-center">
-                    <svg width="28" height="28" viewBox="0 0 28 28" fill="none" xmlns="http://www.w3.org/2000/svg">
-                        <rect x="0.5" y="0.5" width="27" height="27" rx="13.5" fill="white" />
-                        <rect x="0.5" y="0.5" width="27" height="27" rx="13.5" stroke="#D8DAE5" />
-                        <path
-                            opacity="0.2"
-                            d="M13.125 11.375C13.125 11.7211 13.0224 12.0595 12.8301 12.3472C12.6378 12.635 12.3645 12.8593 12.0447 12.9918C11.7249 13.1242 11.3731 13.1589 11.0336 13.0914C10.6941 13.0239 10.3823 12.8572 10.1376 12.6124C9.89282 12.3677 9.72615 12.0559 9.65863 11.7164C9.5911 11.3769 9.62576 11.0251 9.75821 10.7053C9.89066 10.3855 10.115 10.1122 10.4028 9.91993C10.6905 9.72764 11.0289 9.625 11.375 9.625C11.8391 9.625 12.2842 9.80937 12.6124 10.1376C12.9406 10.4658 13.125 10.9109 13.125 11.375ZM16.625 9.625C16.2789 9.625 15.9405 9.72764 15.6528 9.91993C15.365 10.1122 15.1407 10.3855 15.0082 10.7053C14.8758 11.0251 14.8411 11.3769 14.9086 11.7164C14.9762 12.0559 15.1428 12.3677 15.3876 12.6124C15.6323 12.8572 15.9441 13.0239 16.2836 13.0914C16.6231 13.1589 16.9749 13.1242 17.2947 12.9918C17.6145 12.8593 17.8878 12.635 18.0801 12.3472C18.2724 12.0595 18.375 11.7211 18.375 11.375C18.375 10.9109 18.1906 10.4658 17.8624 10.1376C17.5342 9.80937 17.0891 9.625 16.625 9.625ZM11.375 14.875C11.0289 14.875 10.6905 14.9776 10.4028 15.1699C10.115 15.3622 9.89066 15.6355 9.75821 15.9553C9.62576 16.2751 9.5911 16.6269 9.65863 16.9664C9.72615 17.3059 9.89282 17.6177 10.1376 17.8624C10.3823 18.1072 10.6941 18.2739 11.0336 18.3414C11.3731 18.4089 11.7249 18.3742 12.0447 18.2418C12.3645 18.1093 12.6378 17.885 12.8301 17.5972C13.0224 17.3095 13.125 16.9711 13.125 16.625C13.125 16.1609 12.9406 15.7158 12.6124 15.3876C12.2842 15.0594 11.8391 14.875 11.375 14.875Z"
-                            fill="#C1C4D6"
-                        />
-                        <path
-                            d="M11.375 9.1875C10.9424 9.1875 10.5194 9.3158 10.1597 9.55616C9.79996 9.79653 9.51958 10.1382 9.35401 10.5379C9.18845 10.9376 9.14513 11.3774 9.22953 11.8018C9.31394 12.2261 9.52228 12.6159 9.8282 12.9218C10.1341 13.2277 10.5239 13.4361 10.9482 13.5205C11.3726 13.6049 11.8124 13.5616 12.2121 13.396C12.6118 13.2304 12.9535 12.95 13.1938 12.5903C13.4342 12.2306 13.5625 11.8076 13.5625 11.375C13.5625 10.7948 13.332 10.2384 12.9218 9.8282C12.5116 9.41797 11.9552 9.1875 11.375 9.1875ZM11.375 12.6875C11.1154 12.6875 10.8617 12.6105 10.6458 12.4663C10.43 12.3221 10.2617 12.1171 10.1624 11.8773C10.0631 11.6374 10.0371 11.3735 10.0877 11.1189C10.1384 10.8643 10.2634 10.6305 10.4469 10.4469C10.6305 10.2634 10.8643 10.1384 11.1189 10.0877C11.3735 10.0371 11.6374 10.0631 11.8773 10.1624C12.1171 10.2617 12.3221 10.43 12.4663 10.6458C12.6105 10.8617 12.6875 11.1154 12.6875 11.375C12.6875 11.7231 12.5492 12.0569 12.3031 12.3031C12.0569 12.5492 11.7231 12.6875 11.375 12.6875ZM16.625 13.5625C17.0576 13.5625 17.4806 13.4342 17.8403 13.1938C18.2 12.9535 18.4804 12.6118 18.646 12.2121C18.8116 11.8124 18.8549 11.3726 18.7705 10.9482C18.6861 10.5239 18.4777 10.1341 18.1718 9.8282C17.8659 9.52228 17.4761 9.31394 17.0518 9.22953C16.6274 9.14513 16.1876 9.18845 15.7879 9.35401C15.3882 9.51958 15.0465 9.79996 14.8062 10.1597C14.5658 10.5194 14.4375 10.9424 14.4375 11.375C14.4375 11.9552 14.668 12.5116 15.0782 12.9218C15.4884 13.332 16.0448 13.5625 16.625 13.5625ZM16.625 10.0625C16.8846 10.0625 17.1383 10.1395 17.3542 10.2837C17.57 10.4279 17.7383 10.6329 17.8376 10.8727C17.9369 11.1126 17.9629 11.3765 17.9123 11.6311C17.8616 11.8857 17.7366 12.1195 17.5531 12.3031C17.3695 12.4866 17.1357 12.6116 16.8811 12.6623C16.6265 12.7129 16.3626 12.6869 16.1227 12.5876C15.8829 12.4883 15.6779 12.32 15.5337 12.1042C15.3895 11.8883 15.3125 11.6346 15.3125 11.375C15.3125 11.0269 15.4508 10.6931 15.6969 10.4469C15.9431 10.2008 16.2769 10.0625 16.625 10.0625ZM11.375 14.4375C10.9424 14.4375 10.5194 14.5658 10.1597 14.8062C9.79996 15.0465 9.51958 15.3882 9.35401 15.7879C9.18845 16.1876 9.14513 16.6274 9.22953 17.0518C9.31394 17.4761 9.52228 17.8659 9.8282 18.1718C10.1341 18.4777 10.5239 18.6861 10.9482 18.7705C11.3726 18.8549 11.8124 18.8116 12.2121 18.646C12.6118 18.4804 12.9535 18.2 13.1938 17.8403C13.4342 17.4806 13.5625 17.0576 13.5625 16.625C13.5625 16.0448 13.332 15.4884 12.9218 15.0782C12.5116 14.668 11.9552 14.4375 11.375 14.4375ZM11.375 17.9375C11.1154 17.9375 10.8617 17.8605 10.6458 17.7163C10.43 17.5721 10.2617 17.3671 10.1624 17.1273C10.0631 16.8874 10.0371 16.6235 10.0877 16.3689C10.1384 16.1143 10.2634 15.8805 10.4469 15.6969C10.6305 15.5134 10.8643 15.3884 11.1189 15.3377C11.3735 15.2871 11.6374 15.3131 11.8773 15.4124C12.1171 15.5117 12.3221 15.68 12.4663 15.8958C12.6105 16.1117 12.6875 16.3654 12.6875 16.625C12.6875 16.9731 12.5492 17.3069 12.3031 17.5531C12.0569 17.7992 11.7231 17.9375 11.375 17.9375ZM18.8125 16.625C18.8125 16.741 18.7664 16.8523 18.6844 16.9344C18.6023 17.0164 18.491 17.0625 18.375 17.0625H17.0625V18.375C17.0625 18.491 17.0164 18.6023 16.9344 18.6844C16.8523 18.7664 16.741 18.8125 16.625 18.8125C16.509 18.8125 16.3977 18.7664 16.3156 18.6844C16.2336 18.6023 16.1875 18.491 16.1875 18.375V17.0625H14.875C14.759 17.0625 14.6477 17.0164 14.5656 16.9344C14.4836 16.8523 14.4375 16.741 14.4375 16.625C14.4375 16.509 14.4836 16.3977 14.5656 16.3156C14.6477 16.2336 14.759 16.1875 14.875 16.1875H16.1875V14.875C16.1875 14.759 16.2336 14.6477 16.3156 14.5656C16.3977 14.4836 16.509 14.4375 16.625 14.4375C16.741 14.4375 16.8523 14.4836 16.9344 14.5656C17.0164 14.6477 17.0625 14.759 17.0625 14.875V16.1875H18.375C18.491 16.1875 18.6023 16.2336 18.6844 16.3156C18.7664 16.3977 18.8125 16.509 18.8125 16.625Z"
-                            fill="#C1C4D6"
-                        />
-                    </svg>
+                    <SmallUserImgIcon />
                     <div className="text-smEmphasize text-n900">홍길동</div>
                 </div>
