+++ src/components/Card.tsx
@@ -1,4 +1,5 @@
 import Link from "next/link";
+import { LikedIcon, SmallUserImgIcon, ViewIcon } from "./icons/icons";
 
 interface ItemType {
     id: number;
@@ -30,36 +31,14 @@ const Card = ({ item }: { item: ItemType }) => {
 
                 <div className="border-t border-n300 pt-3 flex justify-between items-center">
                     <div className="flex items-center gap-2">
-                        <svg width="28" height="28" viewBox="0 0 28 28" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <rect x="0.5" y="0.5" width="27" height="27" rx="13.5" fill="white" />
-                            <rect x="0.5" y="0.5" width="27" height="27" rx="13.5" stroke="#D8DAE5" />
-                            <path
-                                opacity="0.2"
-                                d="M13.125 11.375C13.125 11.7211 13.0224 12.0595 12.8301 12.3472C12.6378 12.635 12.3645 12.8593 12.0447 12.9918C11.7249 13.1242 11.3731 13.1589 11.0336 13.0914C10.6941 13.0239 10.3823 12.8572 10.1376 12.6124C9.89282 12.3677 9.72615 12.0559 9.65863 11.7164C9.5911 11.3769 9.62576 11.0251 9.75821 10.7053C9.89066 10.3855 10.115 10.1122 10.4028 9.91993C10.6905 9.72764 11.0289 9.625 11.375 9.625C11.8391 9.625 12.2842 9.80937 12.6124 10.1376C12.9406 10.4658 13.125 10.9109 13.125 11.375ZM16.625 9.625C16.2789 9.625 15.9405 9.72764 15.6528 9.91993C15.365 10.1122 15.1407 10.3855 15.0082 10.7053C14.8758 11.0251 14.8411 11.3769 14.9086 11.7164C14.9762 12.0559 15.1428 12.3677 15.3876 12.6124C15.6323 12.8572 15.9441 13.0239 16.2836 13.0914C16.6231 13.1589 16.9749 13.1242 17.2947 12.9918C17.6145 12.8593 17.8878 12.635 18.0801 12.3472C18.2724 12.0595 18.375 11.7211 18.375 11.375C18.375 10.9109 18.1906 10.4658 17.8624 10.1376C17.5342 9.80937 17.0891 9.625 16.625 9.625ZM11.375 14.875C11.0289 14.875 10.6905 14.9776 10.4028 15.1699C10.115 15.3622 9.89066 15.6355 9.75821 15.9553C9.62576 16.2751 9.5911 16.6269 9.65863 16.9664C9.72615 17.3059 9.89282 17.6177 10.1376 17.8624C10.3823 18.1072 10.6941 18.2739 11.0336 18.3414C11.3731 18.4089 11.7249 18.3742 12.0447 18.2418C12.3645 18.1093 12.6378 17.885 12.8301 17.5972C13.0224 17.3095 13.125 16.9711 13.125 16.625C13.125 16.1609 12.9406 15.7158 12.6124 15.3876C12.2842 15.0594 11.8391 14.875 11.375 14.875Z"
-                                fill="#C1C4D6"
-                            />
-                            <path
-                                d="M11.375 9.1875C10.9424 9.1875 10.5194 9.3158 10.1597 9.55616C9.79996 9.79653 9.51958 10.1382 9.35401 10.5379C9.18845 10.9376 9.14513 11.3774 9.22953 11.8018C9.31394 12.2261 9.52228 12.6159 9.8282 12.9218C10.1341 13.2277 10.5239 13.4361 10.9482 13.5205C11.3726 13.6049 11.8124 13.5616 12.2121 13.396C12.6118 13.2304 12.9535 12.95 13.1938 12.5903C13.4342 12.2306 13.5625 11.8076 13.5625 11.375C13.5625 10.7948 13.332 10.2384 12.9218 9.8282C12.5116 9.41797 11.9552 9.1875 11.375 9.1875ZM11.375 12.6875C11.1154 12.6875 10.8617 12.6105 10.6458 12.4663C10.43 12.3221 10.2617 12.1171 10.1624 11.8773C10.0631 11.6374 10.0371 11.3735 10.0877 11.1189C10.1384 10.8643 10.2634 10.6305 10.4469 10.4469C10.6305 10.2634 10.8643 10.1384 11.1189 10.0877C11.3735 10.0371 11.6374 10.0631 11.8773 10.1624C12.1171 10.2617 12.3221 10.43 12.4663 10.6458C12.6105 10.8617 12.6875 11.1154 12.6875 11.375C12.6875 11.7231 12.5492 12.0569 12.3031 12.3031C12.0569 12.5492 11.7231 12.6875 11.375 12.6875ZM16.625 13.5625C17.0576 13.5625 17.4806 13.4342 17.8403 13.1938C18.2 12.9535 18.4804 12.6118 18.646 12.2121C18.8116 11.8124 18.8549 11.3726 18.7705 10.9482C18.6861 10.5239 18.4777 10.1341 18.1718 9.8282C17.8659 9.52228 17.4761 9.31394 17.0518 9.22953C16.6274 9.14513 16.1876 9.18845 15.7879 9.35401C15.3882 9.51958 15.0465 9.79996 14.8062 10.1597C14.5658 10.5194 14.4375 10.9424 14.4375 11.375C14.4375 11.9552 14.668 12.5116 15.0782 12.9218C15.4884 13.332 16.0448 13.5625 16.625 13.5625ZM16.625 10.0625C16.8846 10.0625 17.1383 10.1395 17.3542 10.2837C17.57 10.4279 17.7383 10.6329 17.8376 10.8727C17.9369 11.1126 17.9629 11.3765 17.9123 11.6311C17.8616 11.8857 17.7366 12.1195 17.5531 12.3031C17.3695 12.4866 17.1357 12.6116 16.8811 12.6623C16.6265 12.7129 16.3626 12.6869 16.1227 12.5876C15.8829 12.4883 15.6779 12.32 15.5337 12.1042C15.3895 11.8883 15.3125 11.6346 15.3125 11.375C15.3125 11.0269 15.4508 10.6931 15.6969 10.4469C15.9431 10.2008 16.2769 10.0625 16.625 10.0625ZM11.375 14.4375C10.9424 14.4375 10.5194 14.5658 10.1597 14.8062C9.79996 15.0465 9.51958 15.3882 9.35401 15.7879C9.18845 16.1876 9.14513 16.6274 9.22953 17.0518C9.31394 17.4761 9.52228 17.8659 9.8282 18.1718C10.1341 18.4777 10.5239 18.6861 10.9482 18.7705C11.3726 18.8549 11.8124 18.8116 12.2121 18.646C12.6118 18.4804 12.9535 18.2 13.1938 17.8403C13.4342 17.4806 13.5625 17.0576 13.5625 16.625C13.5625 16.0448 13.332 15.4884 12.9218 15.0782C12.5116 14.668 11.9552 14.4375 11.375 14.4375ZM11.375 17.9375C11.1154 17.9375 10.8617 17.8605 10.6458 17.7163C10.43 17.5721 10.2617 17.3671 10.1624 17.1273C10.0631 16.8874 10.0371 16.6235 10.0877 16.3689C10.1384 16.1143 10.2634 15.8805 10.4469 15.6969C10.6305 15.5134 10.8643 15.3884 11.1189 15.3377C11.3735 15.2871 11.6374 15.3131 11.8773 15.4124C12.1171 15.5117 12.3221 15.68 12.4663 15.8958C12.6105 16.1117 12.6875 16.3654 12.6875 16.625C12.6875 16.9731 12.5492 17.3069 12.3031 17.5531C12.0569 17.7992 11.7231 17.9375 11.375 17.9375ZM18.8125 16.625C18.8125 16.741 18.7664 16.8523 18.6844 16.9344C18.6023 17.0164 18.491 17.0625 18.375 17.0625H17.0625V18.375C17.0625 18.491 17.0164 18.6023 16.9344 18.6844C16.8523 18.7664 16.741 18.8125 16.625 18.8125C16.509 18.8125 16.3977 18.7664 16.3156 18.6844C16.2336 18.6023 16.1875 18.491 16.1875 18.375V17.0625H14.875C14.759 17.0625 14.6477 17.0164 14.5656 16.9344C14.4836 16.8523 14.4375 16.741 14.4375 16.625C14.4375 16.509 14.4836 16.3977 14.5656 16.3156C14.6477 16.2336 14.759 16.1875 14.875 16.1875H16.1875V14.875C16.1875 14.759 16.2336 14.6477 16.3156 14.5656C16.3977 14.4836 16.509 14.4375 16.625 14.4375C16.741 14.4375 16.8523 14.4836 16.9344 14.5656C17.0164 14.6477 17.0625 14.759 17.0625 14.875V16.1875H18.375C18.491 16.1875 18.6023 16.2336 18.6844 16.3156C18.7664 16.3977 18.8125 16.509 18.8125 16.625Z"
-                                fill="#C1C4D6"
-                            />
-                        </svg>
+                        <SmallUserImgIcon />
 
                         <div className="text-smEmphasize text-n900">{item.creator}</div>
                     </div>
                     <div className="flex items-center">
-                        <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <path
-                                d="M13.9062 3.125C12.293 3.125 10.8805 3.81875 10 4.99141C9.11953 3.81875 7.70703 3.125 6.09375 3.125C4.80955 3.12645 3.57837 3.63723 2.6703 4.5453C1.76223 5.45337 1.25145 6.68455 1.25 7.96875C1.25 13.4375 9.35859 17.8641 9.70391 18.0469C9.79492 18.0958 9.89665 18.1215 10 18.1215C10.1033 18.1215 10.2051 18.0958 10.2961 18.0469C10.6414 17.8641 18.75 13.4375 18.75 7.96875C18.7486 6.68455 18.2378 5.45337 17.3297 4.5453C16.4216 3.63723 15.1904 3.12645 13.9062 3.125ZM10 16.7813C8.57344 15.95 2.5 12.1633 2.5 7.96875C2.50124 7.01601 2.88026 6.10265 3.55396 5.42896C4.22765 4.75526 5.14101 4.37624 6.09375 4.375C7.61328 4.375 8.88906 5.18438 9.42188 6.48438C9.46896 6.59901 9.54907 6.69705 9.65201 6.76605C9.75494 6.83505 9.87607 6.8719 10 6.8719C10.1239 6.8719 10.2451 6.83505 10.348 6.76605C10.4509 6.69705 10.531 6.59901 10.5781 6.48438C11.1109 5.18203 12.3867 4.375 13.9062 4.375C14.859 4.37624 15.7724 4.75526 16.446 5.42896C17.1197 6.10265 17.4988 7.01601 17.5 7.96875C17.5 12.157 11.425 15.9492 10 16.7813Z"
-                                fill="#FF6363"
-                            />
-                        </svg>
+                        <LikedIcon />
                         <div className="text-sm text-black ml-1.5 mr-3">{item.likeCount}</div>
-                        <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <path
-                                d="M19.3211 9.74688C19.2937 9.68516 18.632 8.21719 17.1609 6.74609C15.2008 4.78594 12.725 3.75 9.99999 3.75C7.27499 3.75 4.79921 4.78594 2.83905 6.74609C1.36796 8.21719 0.703118 9.6875 0.678899 9.74688C0.643362 9.82681 0.625 9.91331 0.625 10.0008C0.625 10.0883 0.643362 10.1748 0.678899 10.2547C0.706243 10.3164 1.36796 11.7836 2.83905 13.2547C4.79921 15.2141 7.27499 16.25 9.99999 16.25C12.725 16.25 15.2008 15.2141 17.1609 13.2547C18.632 11.7836 19.2937 10.3164 19.3211 10.2547C19.3566 10.1748 19.375 10.0883 19.375 10.0008C19.375 9.91331 19.3566 9.82681 19.3211 9.74688ZM9.99999 15C7.5953 15 5.49452 14.1258 3.75546 12.4023C3.0419 11.6927 2.43483 10.8836 1.95312 10C2.4347 9.11636 3.04179 8.30717 3.75546 7.59766C5.49452 5.87422 7.5953 5 9.99999 5C12.4047 5 14.5055 5.87422 16.2445 7.59766C16.9595 8.307 17.5679 9.11619 18.0508 10C17.4875 11.0516 15.0336 15 9.99999 15ZM9.99999 6.25C9.25831 6.25 8.53329 6.46993 7.9166 6.88199C7.29992 7.29404 6.81927 7.87971 6.53544 8.56494C6.25162 9.25016 6.17735 10.0042 6.32205 10.7316C6.46674 11.459 6.82389 12.1272 7.34834 12.6517C7.87279 13.1761 8.54097 13.5333 9.2684 13.6779C9.99583 13.8226 10.7498 13.7484 11.4351 13.4645C12.1203 13.1807 12.7059 12.7001 13.118 12.0834C13.5301 11.4667 13.75 10.7417 13.75 10C13.749 9.00576 13.3535 8.05253 12.6505 7.34949C11.9475 6.64645 10.9942 6.25103 9.99999 6.25ZM9.99999 12.5C9.50554 12.5 9.02219 12.3534 8.61107 12.0787C8.19994 11.804 7.87951 11.4135 7.69029 10.9567C7.50107 10.4999 7.45157 9.99723 7.54803 9.51227C7.64449 9.02732 7.88259 8.58186 8.23222 8.23223C8.58186 7.8826 9.02731 7.6445 9.51227 7.54804C9.99722 7.45157 10.4999 7.50108 10.9567 7.6903C11.4135 7.87952 11.804 8.19995 12.0787 8.61107C12.3534 9.0222 12.5 9.50555 12.5 10C12.5 10.663 12.2366 11.2989 11.7678 11.7678C11.2989 12.2366 10.663 12.5 9.99999 12.5Z"
-                                fill="#474D66"
-                            />
-                        </svg>
+                        <ViewIcon />
                         <div className="text-sm text-black ml-1.5">{item.viewCount}</div>
                     </div>
                 </div>
