+++ src/components/gnb/Nav.tsx
@@ -4,8 +4,9 @@ import Logo from "@/assets/logo.png";
 import Image from "next/image";
 import Link from "next/link";
 import { useState } from "react";
+import { FolderIcon, HomeIcon, NotificationIcon, ProfileIcon } from "../icons/icons";
 import FindPwModal from "./FindPwModal";
-import LoginModal from "./LoginModal";
+import LoginModal from "./LoginModal/LoginModal";
 
 export default function Nav() {
     const [isShowLoginModal, setIsShowLoginModal] = useState(false);
@@ -41,40 +42,3 @@ export default function Nav() {
         </nav>
     );
 }
-
-const HomeIcon = () => (
-    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
-        <path
-            d="M20.5603 10.1887L13.0603 2.68875C12.779 2.40766 12.3977 2.24976 12 2.24976C11.6024 2.24976 11.221 2.40766 10.9397 2.68875L3.4397 10.1887C3.29973 10.3277 3.18878 10.493 3.1133 10.6752C3.03782 10.8574 2.99931 11.0528 3.00001 11.25V20.25C3.00001 20.4489 3.07903 20.6397 3.21968 20.7803C3.36033 20.921 3.5511 21 3.75001 21H9.75001C9.94892 21 10.1397 20.921 10.2803 20.7803C10.421 20.6397 10.5 20.4489 10.5 20.25V15H13.5V20.25C13.5 20.4489 13.579 20.6397 13.7197 20.7803C13.8603 20.921 14.0511 21 14.25 21H20.25C20.4489 21 20.6397 20.921 20.7803 20.7803C20.921 20.6397 21 20.4489 21 20.25V11.25C21.0007 11.0528 20.9622 10.8574 20.8867 10.6752C20.8112 10.493 20.7003 10.3277 20.5603 10.1887ZM19.5 19.5H15V14.25C15 14.0511 14.921 13.8603 14.7803 13.7197C14.6397 13.579 14.4489 13.5 14.25 13.5H9.75001C9.5511 13.5 9.36033 13.579 9.21968 13.7197C9.07903 13.8603 9.00001 14.0511 9.00001 14.25V19.5H4.50001V11.25L12 3.75L19.5 11.25V19.5Z"
-            fill="#474D66"
-        />
-    </svg>
-);
-
-const FolderIcon = () => (
-    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
-        <path
-            d="M20.25 6.75H12.3103L9.75 4.18969C9.61122 4.04979 9.44601 3.93888 9.26398 3.8634C9.08196 3.78792 8.88674 3.74937 8.68969 3.75H3.75C3.35218 3.75 2.97064 3.90804 2.68934 4.18934C2.40804 4.47064 2.25 4.85218 2.25 5.25V18.8081C2.2505 19.1904 2.40257 19.5568 2.67286 19.8271C2.94316 20.0974 3.30962 20.2495 3.69188 20.25H20.3334C20.709 20.2495 21.069 20.1001 21.3346 19.8346C21.6001 19.569 21.7495 19.209 21.75 18.8334V8.25C21.75 7.85218 21.592 7.47064 21.3107 7.18934C21.0294 6.90804 20.6478 6.75 20.25 6.75ZM3.75 5.25H8.68969L10.1897 6.75H3.75V5.25ZM20.25 18.75H3.75V8.25H20.25V18.75Z"
-            fill="#474D66"
-        />
-    </svg>
-);
-
-const NotificationIcon = () => (
-    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
-        <path
-            d="M20.7946 16.4944C20.2743 15.5981 19.5009 13.0622 19.5009 9.75C19.5009 7.76088 18.7107 5.85322 17.3042 4.4467C15.8976 3.04018 13.99 2.25 12.0009 2.25C10.0117 2.25 8.10409 3.04018 6.69757 4.4467C5.29104 5.85322 4.50087 7.76088 4.50087 9.75C4.50087 13.0631 3.72649 15.5981 3.20618 16.4944C3.07331 16.7222 3.00287 16.9811 3.00196 17.2449C3.00106 17.5086 3.06972 17.768 3.20102 17.9967C3.33233 18.2255 3.52163 18.4156 3.74984 18.5478C3.97805 18.6801 4.2371 18.7498 4.50087 18.75H8.3268C8.49984 19.5967 8.96001 20.3577 9.62949 20.9042C10.299 21.4507 11.1367 21.7492 12.0009 21.7492C12.8651 21.7492 13.7028 21.4507 14.3722 20.9042C15.0417 20.3577 15.5019 19.5967 15.6749 18.75H19.5009C19.7645 18.7496 20.0235 18.6798 20.2516 18.5475C20.4796 18.4151 20.6688 18.225 20.8 17.9963C20.9312 17.7676 20.9998 17.5083 20.9988 17.2446C20.9979 16.9809 20.9274 16.7222 20.7946 16.4944ZM12.0009 20.25C11.5357 20.2499 11.082 20.1055 10.7022 19.8369C10.3225 19.5683 10.0353 19.1886 9.88024 18.75H14.1215C13.9664 19.1886 13.6793 19.5683 13.2995 19.8369C12.9197 20.1055 12.466 20.2499 12.0009 20.25ZM4.50087 17.25C5.22274 16.0087 6.00087 13.1325 6.00087 9.75C6.00087 8.1587 6.63301 6.63258 7.75823 5.50736C8.88344 4.38214 10.4096 3.75 12.0009 3.75C13.5922 3.75 15.1183 4.38214 16.2435 5.50736C17.3687 6.63258 18.0009 8.1587 18.0009 9.75C18.0009 13.1297 18.7771 16.0059 19.5009 17.25H4.50087Z"
-            fill="#474D66"
-        />
-        <circle cx="21.5" cy="2.5" r="2.5" fill="#FF6363" />
-    </svg>
-);
-
-const ProfileIcon = () => (
-    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
-        <path
-            d="M12 2.25C10.0716 2.25 8.18657 2.82183 6.58319 3.89317C4.97982 4.96451 3.73013 6.48726 2.99218 8.26884C2.25422 10.0504 2.06114 12.0108 2.43735 13.9021C2.81355 15.7934 3.74215 17.5307 5.10571 18.8943C6.46928 20.2579 8.20656 21.1865 10.0979 21.5627C11.9892 21.9389 13.9496 21.7458 15.7312 21.0078C17.5127 20.2699 19.0355 19.0202 20.1068 17.4168C21.1782 15.8134 21.75 13.9284 21.75 12C21.7473 9.41498 20.7192 6.93661 18.8913 5.10872C17.0634 3.28084 14.585 2.25273 12 2.25ZM6.945 18.5156C7.48757 17.6671 8.23501 16.9688 9.11843 16.4851C10.0019 16.0013 10.9928 15.7478 12 15.7478C13.0072 15.7478 13.9982 16.0013 14.8816 16.4851C15.765 16.9688 16.5124 17.6671 17.055 18.5156C15.6097 19.6397 13.831 20.2499 12 20.2499C10.169 20.2499 8.39032 19.6397 6.945 18.5156ZM9 11.25C9 10.6567 9.17595 10.0766 9.5056 9.58329C9.83524 9.08994 10.3038 8.70542 10.852 8.47836C11.4001 8.2513 12.0033 8.19189 12.5853 8.30764C13.1672 8.4234 13.7018 8.70912 14.1213 9.12868C14.5409 9.54824 14.8266 10.0828 14.9424 10.6647C15.0581 11.2467 14.9987 11.8499 14.7716 12.3981C14.5446 12.9462 14.1601 13.4148 13.6667 13.7444C13.1734 14.0741 12.5933 14.25 12 14.25C11.2044 14.25 10.4413 13.9339 9.87868 13.3713C9.31607 12.8087 9 12.0456 9 11.25ZM18.165 17.4759C17.3285 16.2638 16.1524 15.3261 14.7844 14.7806C15.5192 14.2019 16.0554 13.4085 16.3184 12.5108C16.5815 11.6132 16.5582 10.6559 16.252 9.77207C15.9457 8.88825 15.3716 8.12183 14.6096 7.5794C13.8475 7.03696 12.9354 6.74548 12 6.74548C11.0646 6.74548 10.1525 7.03696 9.39044 7.5794C8.62839 8.12183 8.05432 8.88825 7.74805 9.77207C7.44179 10.6559 7.41855 11.6132 7.68157 12.5108C7.94459 13.4085 8.4808 14.2019 9.21563 14.7806C7.84765 15.3261 6.67147 16.2638 5.835 17.4759C4.77804 16.2873 4.0872 14.8185 3.84567 13.2464C3.60415 11.6743 3.82224 10.0658 4.47368 8.61478C5.12512 7.16372 6.18213 5.93192 7.51745 5.06769C8.85276 4.20346 10.4094 3.74367 12 3.74367C13.5906 3.74367 15.1473 4.20346 16.4826 5.06769C17.8179 5.93192 18.8749 7.16372 19.5263 8.61478C20.1778 10.0658 20.3959 11.6743 20.1543 13.2464C19.9128 14.8185 19.222 16.2873 18.165 17.4759Z"
-            fill="#474D66"
-        />
-    </svg>
-);
