+++ src/components/icons/icons.tsx
@@ -1,0 +1,245 @@
+export const HomeIcon = () => (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
+        <path
+            d="M20.5603 10.1887L13.0603 2.68875C12.779 2.40766 12.3977 2.24976 12 2.24976C11.6024 2.24976 11.221 2.40766 10.9397 2.68875L3.4397 10.1887C3.29973 10.3277 3.18878 10.493 3.1133 10.6752C3.03782 10.8574 2.99931 11.0528 3.00001 11.25V20.25C3.00001 20.4489 3.07903 20.6397 3.21968 20.7803C3.36033 20.921 3.5511 21 3.75001 21H9.75001C9.94892 21 10.1397 20.921 10.2803 20.7803C10.421 20.6397 10.5 20.4489 10.5 20.25V15H13.5V20.25C13.5 20.4489 13.579 20.6397 13.7197 20.7803C13.8603 20.921 14.0511 21 14.25 21H20.25C20.4489 21 20.6397 20.921 20.7803 20.7803C20.921 20.6397 21 20.4489 21 20.25V11.25C21.0007 11.0528 20.9622 10.8574 20.8867 10.6752C20.8112 10.493 20.7003 10.3277 20.5603 10.1887ZM19.5 19.5H15V14.25C15 14.0511 14.921 13.8603 14.7803 13.7197C14.6397 13.579 14.4489 13.5 14.25 13.5H9.75001C9.5511 13.5 9.36033 13.579 9.21968 13.7197C9.07903 13.8603 9.00001 14.0511 9.00001 14.25V19.5H4.50001V11.25L12 3.75L19.5 11.25V19.5Z"
+            fill="#474D66"
+        />
+    </svg>
+);
+
+export const FolderIcon = () => (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
+        <path
+            d="M20.25 6.75H12.3103L9.75 4.18969C9.61122 4.04979 9.44601 3.93888 9.26398 3.8634C9.08196 3.78792 8.88674 3.74937 8.68969 3.75H3.75C3.35218 3.75 2.97064 3.90804 2.68934 4.18934C2.40804 4.47064 2.25 4.85218 2.25 5.25V18.8081C2.2505 19.1904 2.40257 19.5568 2.67286 19.8271C2.94316 20.0974 3.30962 20.2495 3.69188 20.25H20.3334C20.709 20.2495 21.069 20.1001 21.3346 19.8346C21.6001 19.569 21.7495 19.209 21.75 18.8334V8.25C21.75 7.85218 21.592 7.47064 21.3107 7.18934C21.0294 6.90804 20.6478 6.75 20.25 6.75ZM3.75 5.25H8.68969L10.1897 6.75H3.75V5.25ZM20.25 18.75H3.75V8.25H20.25V18.75Z"
+            fill="#474D66"
+        />
+    </svg>
+);
+
+export const NotificationIcon = () => (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
+        <path
+            d="M20.7946 16.4944C20.2743 15.5981 19.5009 13.0622 19.5009 9.75C19.5009 7.76088 18.7107 5.85322 17.3042 4.4467C15.8976 3.04018 13.99 2.25 12.0009 2.25C10.0117 2.25 8.10409 3.04018 6.69757 4.4467C5.29104 5.85322 4.50087 7.76088 4.50087 9.75C4.50087 13.0631 3.72649 15.5981 3.20618 16.4944C3.07331 16.7222 3.00287 16.9811 3.00196 17.2449C3.00106 17.5086 3.06972 17.768 3.20102 17.9967C3.33233 18.2255 3.52163 18.4156 3.74984 18.5478C3.97805 18.6801 4.2371 18.7498 4.50087 18.75H8.3268C8.49984 19.5967 8.96001 20.3577 9.62949 20.9042C10.299 21.4507 11.1367 21.7492 12.0009 21.7492C12.8651 21.7492 13.7028 21.4507 14.3722 20.9042C15.0417 20.3577 15.5019 19.5967 15.6749 18.75H19.5009C19.7645 18.7496 20.0235 18.6798 20.2516 18.5475C20.4796 18.4151 20.6688 18.225 20.8 17.9963C20.9312 17.7676 20.9998 17.5083 20.9988 17.2446C20.9979 16.9809 20.9274 16.7222 20.7946 16.4944ZM12.0009 20.25C11.5357 20.2499 11.082 20.1055 10.7022 19.8369C10.3225 19.5683 10.0353 19.1886 9.88024 18.75H14.1215C13.9664 19.1886 13.6793 19.5683 13.2995 19.8369C12.9197 20.1055 12.466 20.2499 12.0009 20.25ZM4.50087 17.25C5.22274 16.0087 6.00087 13.1325 6.00087 9.75C6.00087 8.1587 6.63301 6.63258 7.75823 5.50736C8.88344 4.38214 10.4096 3.75 12.0009 3.75C13.5922 3.75 15.1183 4.38214 16.2435 5.50736C17.3687 6.63258 18.0009 8.1587 18.0009 9.75C18.0009 13.1297 18.7771 16.0059 19.5009 17.25H4.50087Z"
+            fill="#474D66"
+        />
+        <circle cx="21.5" cy="2.5" r="2.5" fill="#FF6363" />
+    </svg>
+);
+
+export const ProfileIcon = () => (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
+        <path
+            d="M12 2.25C10.0716 2.25 8.18657 2.82183 6.58319 3.89317C4.97982 4.96451 3.73013 6.48726 2.99218 8.26884C2.25422 10.0504 2.06114 12.0108 2.43735 13.9021C2.81355 15.7934 3.74215 17.5307 5.10571 18.8943C6.46928 20.2579 8.20656 21.1865 10.0979 21.5627C11.9892 21.9389 13.9496 21.7458 15.7312 21.0078C17.5127 20.2699 19.0355 19.0202 20.1068 17.4168C21.1782 15.8134 21.75 13.9284 21.75 12C21.7473 9.41498 20.7192 6.93661 18.8913 5.10872C17.0634 3.28084 14.585 2.25273 12 2.25ZM6.945 18.5156C7.48757 17.6671 8.23501 16.9688 9.11843 16.4851C10.0019 16.0013 10.9928 15.7478 12 15.7478C13.0072 15.7478 13.9982 16.0013 14.8816 16.4851C15.765 16.9688 16.5124 17.6671 17.055 18.5156C15.6097 19.6397 13.831 20.2499 12 20.2499C10.169 20.2499 8.39032 19.6397 6.945 18.5156ZM9 11.25C9 10.6567 9.17595 10.0766 9.5056 9.58329C9.83524 9.08994 10.3038 8.70542 10.852 8.47836C11.4001 8.2513 12.0033 8.19189 12.5853 8.30764C13.1672 8.4234 13.7018 8.70912 14.1213 9.12868C14.5409 9.54824 14.8266 10.0828 14.9424 10.6647C15.0581 11.2467 14.9987 11.8499 14.7716 12.3981C14.5446 12.9462 14.1601 13.4148 13.6667 13.7444C13.1734 14.0741 12.5933 14.25 12 14.25C11.2044 14.25 10.4413 13.9339 9.87868 13.3713C9.31607 12.8087 9 12.0456 9 11.25ZM18.165 17.4759C17.3285 16.2638 16.1524 15.3261 14.7844 14.7806C15.5192 14.2019 16.0554 13.4085 16.3184 12.5108C16.5815 11.6132 16.5582 10.6559 16.252 9.77207C15.9457 8.88825 15.3716 8.12183 14.6096 7.5794C13.8475 7.03696 12.9354 6.74548 12 6.74548C11.0646 6.74548 10.1525 7.03696 9.39044 7.5794C8.62839 8.12183 8.05432 8.88825 7.74805 9.77207C7.44179 10.6559 7.41855 11.6132 7.68157 12.5108C7.94459 13.4085 8.4808 14.2019 9.21563 14.7806C7.84765 15.3261 6.67147 16.2638 5.835 17.4759C4.77804 16.2873 4.0872 14.8185 3.84567 13.2464C3.60415 11.6743 3.82224 10.0658 4.47368 8.61478C5.12512 7.16372 6.18213 5.93192 7.51745 5.06769C8.85276 4.20346 10.4094 3.74367 12 3.74367C13.5906 3.74367 15.1473 4.20346 16.4826 5.06769C17.8179 5.93192 18.8749 7.16372 19.5263 8.61478C20.1778 10.0658 20.3959 11.6743 20.1543 13.2464C19.9128 14.8185 19.222 16.2873 18.165 17.4759Z"
+            fill="#474D66"
+        />
+    </svg>
+);
+
+export const GoogleIcon = () => (
+    <svg width="25" height="25" viewBox="0 0 25 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M24.02 12.7727C24.02 11.9218 23.9436 11.1036 23.8018 10.3181H12.5V14.9599H18.9582C18.68 16.4599 17.8345 17.7308 16.5636 18.5818V21.5927H20.4418C22.7109 19.5036 24.02 16.4272 24.02 12.7727Z"
+            fill="#4285F4"
+        />
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M12.5 24.5C15.74 24.5 18.4564 23.4255 20.4418 21.5928L16.5636 18.5818C15.4891 19.3018 14.1145 19.7273 12.5 19.7273C9.37455 19.7273 6.72909 17.6164 5.78546 14.78H1.77637V17.8891C3.75091 21.8109 7.80909 24.5 12.5 24.5Z"
+            fill="#34A853"
+        />
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M5.78545 14.7799C5.54545 14.0599 5.40909 13.2908 5.40909 12.4999C5.40909 11.709 5.54545 10.9399 5.78545 10.2199V7.11084H1.77636C0.963636 8.73084 0.5 10.5636 0.5 12.4999C0.5 14.4363 0.963636 16.269 1.77636 17.889L5.78545 14.7799Z"
+            fill="#FBBC05"
+        />
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M12.5 5.27273C14.2618 5.27273 15.8436 5.87818 17.0873 7.06727L20.5291 3.62545C18.4509 1.68909 15.7345 0.5 12.5 0.5C7.80909 0.5 3.75091 3.18909 1.77637 7.11091L5.78546 10.22C6.72909 7.38364 9.37455 5.27273 12.5 5.27273Z"
+            fill="#EA4335"
+        />
+    </svg>
+);
+
+export const GithubIcon = () => (
+    <svg width="24" height="25" viewBox="0 0 24 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <g clipPath="url(#clip0_678_2)">
+            <path
+                fillRule="evenodd"
+                clipRule="evenodd"
+                d="M11.9642 0.5C5.34833 0.5 0 6 0 12.8042C0 18.2432 3.42686 22.8472 8.18082 24.4767C8.77518 24.5992 8.9929 24.212 8.9929 23.8862C8.9929 23.601 8.97331 22.6232 8.97331 21.6045C5.64514 22.338 4.95208 20.1378 4.95208 20.1378C4.41722 18.7118 3.62473 18.3452 3.62473 18.3452C2.53543 17.5915 3.70408 17.5915 3.70408 17.5915C4.91241 17.673 5.54645 18.8545 5.54645 18.8545C6.61592 20.7285 8.33926 20.199 9.03257 19.873C9.13151 19.0785 9.44865 18.5285 9.78539 18.223C7.13094 17.9377 4.33812 16.8785 4.33812 12.1523C4.33812 10.8078 4.81322 9.70775 5.56604 8.85225C5.44727 8.54675 5.03118 7.2835 5.68506 5.59275C5.68506 5.59275 6.69527 5.26675 8.97306 6.85575C9.94827 6.58642 10.954 6.4494 11.9642 6.44825C12.9744 6.44825 14.0042 6.591 14.9552 6.85575C17.2332 5.26675 18.2434 5.59275 18.2434 5.59275C18.8973 7.2835 18.481 8.54675 18.3622 8.85225C19.1349 9.70775 19.5904 10.8078 19.5904 12.1523C19.5904 16.8785 16.7976 17.9172 14.1233 18.223C14.5592 18.61 14.9353 19.3433 14.9353 20.5045C14.9353 22.1545 14.9158 23.4788 14.9158 23.886C14.9158 24.212 15.1337 24.5992 15.7278 24.477C20.4818 22.847 23.9087 18.2432 23.9087 12.8042C23.9282 6 18.5603 0.5 11.9642 0.5Z"
+                fill="white"
+            />
+        </g>
+        <defs>
+            <clipPath id="clip0_678_2">
+                <rect width="24" height="24" fill="white" transform="translate(0 0.5)" />
+            </clipPath>
+        </defs>
+    </svg>
+);
+
+export const KakaoIcon = () => (
+    <svg width="22" height="23" viewBox="0 0 22 23" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path
+            opacity="0.902"
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M11 1.39062C5.24333 1.39062 0 6.12079 0 10.1226C0 13.1211 1.90422 15.766 4.80456 17.3377L3.58478 21.918C3.476 22.324 3.927 22.6464 4.27289 22.4127L9.62256 18.7833C10.0736 18.8283 10.5331 18.8545 11 18.8545C17.0744 18.8545 22 14.9452 22 10.1226C22 6.12079 17.0744 1.39062 11 1.39062Z"
+            fill="black"
+        />
+    </svg>
+);
+
+export const CloseBtn = ({ onClose }: { onClose: () => void }) => {
+    return (
+        <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer" onClick={onClose}>
+            <path
+                d="M19.281 18.2193C19.3507 18.289 19.406 18.3717 19.4437 18.4628C19.4814 18.5538 19.5008 18.6514 19.5008 18.7499C19.5008 18.8485 19.4814 18.9461 19.4437 19.0371C19.406 19.1281 19.3507 19.2109 19.281 19.2806C19.2114 19.3502 19.1286 19.4055 19.0376 19.4432C18.9465 19.4809 18.849 19.5003 18.7504 19.5003C18.6519 19.5003 18.5543 19.4809 18.4632 19.4432C18.3722 19.4055 18.2895 19.3502 18.2198 19.2806L12.0004 13.0602L5.78104 19.2806C5.64031 19.4213 5.44944 19.5003 5.25042 19.5003C5.05139 19.5003 4.86052 19.4213 4.71979 19.2806C4.57906 19.1398 4.5 18.949 4.5 18.7499C4.5 18.5509 4.57906 18.36 4.71979 18.2193L10.9401 11.9999L4.71979 5.78055C4.57906 5.63982 4.5 5.44895 4.5 5.24993C4.5 5.05091 4.57906 4.86003 4.71979 4.7193C4.86052 4.57857 5.05139 4.49951 5.25042 4.49951C5.44944 4.49951 5.64031 4.57857 5.78104 4.7193L12.0004 10.9396L18.2198 4.7193C18.3605 4.57857 18.5514 4.49951 18.7504 4.49951C18.9494 4.49951 19.1403 4.57857 19.281 4.7193C19.4218 4.86003 19.5008 5.05091 19.5008 5.24993C19.5008 5.44895 19.4218 5.63982 19.281 5.78055L13.0607 11.9999L19.281 18.2193Z"
+                fill="#474D66"
+            />
+        </svg>
+    );
+};
+
+export const CheckIcon = () => {
+    return (
+        <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+                d="M10 0.25C8.07164 0.25 6.18657 0.821828 4.58319 1.89317C2.97982 2.96451 1.73013 4.48726 0.992179 6.26884C0.254225 8.05042 0.061142 10.0108 0.437348 11.9021C0.813554 13.7934 1.74215 15.5307 3.10571 16.8943C4.46928 18.2579 6.20656 19.1865 8.09787 19.5627C9.98919 19.9389 11.9496 19.7458 13.7312 19.0078C15.5127 18.2699 17.0355 17.0202 18.1068 15.4168C19.1782 13.8134 19.75 11.9284 19.75 10C19.7473 7.41498 18.7192 4.93661 16.8913 3.10872C15.0634 1.28084 12.585 0.25273 10 0.25ZM14.2806 8.28063L9.03063 13.5306C8.96097 13.6004 8.87826 13.6557 8.78721 13.6934C8.69616 13.7312 8.59857 13.7506 8.5 13.7506C8.40144 13.7506 8.30385 13.7312 8.2128 13.6934C8.12175 13.6557 8.03903 13.6004 7.96938 13.5306L5.71938 11.2806C5.57865 11.1399 5.49959 10.949 5.49959 10.75C5.49959 10.551 5.57865 10.3601 5.71938 10.2194C5.86011 10.0786 6.05098 9.99958 6.25 9.99958C6.44903 9.99958 6.6399 10.0786 6.78063 10.2194L8.5 11.9397L13.2194 7.21937C13.2891 7.14969 13.3718 7.09442 13.4628 7.0567C13.5539 7.01899 13.6515 6.99958 13.75 6.99958C13.8486 6.99958 13.9461 7.01899 14.0372 7.0567C14.1282 7.09442 14.2109 7.14969 14.2806 7.21937C14.3503 7.28906 14.4056 7.37178 14.4433 7.46283C14.481 7.55387 14.5004 7.65145 14.5004 7.75C14.5004 7.84855 14.481 7.94613 14.4433 8.03717C14.4056 8.12822 14.3503 8.21094 14.2806 8.28063Z"
+                fill="#FF6363"
+            />
+        </svg>
+    );
+};
+
+export const NotCheckIcon = () => {
+    return <div className="w-5 h-5 rounded-full border border-[#8F95B2]"></div>;
+};
+
+export const SmallCheckIcon = () => {
+    return (
+        <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+                d="M10 0.25C8.07164 0.25 6.18657 0.821828 4.58319 1.89317C2.97982 2.96451 1.73013 4.48726 0.992179 6.26884C0.254225 8.05042 0.061142 10.0108 0.437348 11.9021C0.813554 13.7934 1.74215 15.5307 3.10571 16.8943C4.46928 18.2579 6.20656 19.1865 8.09787 19.5627C9.98919 19.9389 11.9496 19.7458 13.7312 19.0078C15.5127 18.2699 17.0355 17.0202 18.1068 15.4168C19.1782 13.8134 19.75 11.9284 19.75 10C19.7473 7.41498 18.7192 4.93661 16.8913 3.10872C15.0634 1.28084 12.585 0.25273 10 0.25ZM14.2806 8.28063L9.03063 13.5306C8.96097 13.6004 8.87826 13.6557 8.78721 13.6934C8.69616 13.7312 8.59857 13.7506 8.5 13.7506C8.40144 13.7506 8.30385 13.7312 8.2128 13.6934C8.12175 13.6557 8.03903 13.6004 7.96938 13.5306L5.71938 11.2806C5.57865 11.1399 5.49959 10.949 5.49959 10.75C5.49959 10.551 5.57865 10.3601 5.71938 10.2194C5.86011 10.0786 6.05098 9.99958 6.25 9.99958C6.44903 9.99958 6.6399 10.0786 6.78063 10.2194L8.5 11.9397L13.2194 7.21937C13.2891 7.14969 13.3718 7.09442 13.4628 7.0567C13.5539 7.01899 13.6515 6.99958 13.75 6.99958C13.8486 6.99958 13.9461 7.01899 14.0372 7.0567C14.1282 7.09442 14.2109 7.14969 14.2806 7.21937C14.3503 7.28906 14.4056 7.37178 14.4433 7.46283C14.481 7.55387 14.5004 7.65145 14.5004 7.75C14.5004 7.84855 14.481 7.94613 14.4433 8.03717C14.4056 8.12822 14.3503 8.21094 14.2806 8.28063Z"
+                fill="#FF6363"
+            />
+        </svg>
+    );
+};
+
+export const SmallNotCheckIcon = () => {
+    return (
+        <svg width="18" height="18" viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+                d="M9 0.875C7.39303 0.875 5.82214 1.35152 4.486 2.24431C3.14985 3.1371 2.10844 4.40605 1.49348 5.8907C0.87852 7.37535 0.717618 9.00901 1.03112 10.5851C1.34463 12.1612 2.11846 13.6089 3.25476 14.7452C4.39106 15.8815 5.8388 16.6554 7.4149 16.9689C8.99099 17.2824 10.6247 17.1215 12.1093 16.5065C13.594 15.8916 14.8629 14.8502 15.7557 13.514C16.6485 12.1779 17.125 10.607 17.125 9C17.1227 6.84581 16.266 4.78051 14.7427 3.25727C13.2195 1.73403 11.1542 0.877275 9 0.875ZM9 15.875C7.64026 15.875 6.31105 15.4718 5.18046 14.7164C4.04987 13.9609 3.16868 12.8872 2.64833 11.6309C2.12798 10.3747 1.99183 8.99237 2.2571 7.65875C2.52238 6.32513 3.17716 5.10013 4.13864 4.13864C5.10013 3.17716 6.32514 2.52237 7.65876 2.2571C8.99238 1.99183 10.3747 2.12798 11.631 2.64833C12.8872 3.16868 13.9609 4.04987 14.7164 5.18045C15.4718 6.31104 15.875 7.64025 15.875 9C15.8729 10.8227 15.1479 12.5702 13.8591 13.8591C12.5702 15.1479 10.8227 15.8729 9 15.875Z"
+                fill="#8F95B2"
+            />
+        </svg>
+    );
+};
+
+export const ShareIcon = () => {
+    return (
+        <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+                d="M12.9422 7.05776C13.0003 7.11581 13.0464 7.18474 13.0778 7.26061C13.1093 7.33648 13.1255 7.41781 13.1255 7.49995C13.1255 7.58208 13.1093 7.66341 13.0778 7.73929C13.0464 7.81516 13.0003 7.88409 12.9422 7.94214L7.94217 12.9421C7.8841 13.0002 7.81516 13.0463 7.73929 13.0777C7.66342 13.1091 7.5821 13.1253 7.49998 13.1253C7.41786 13.1253 7.33654 13.1091 7.26067 13.0777C7.1848 13.0463 7.11586 13.0002 7.05779 12.9421C6.99972 12.8841 6.95366 12.8151 6.92223 12.7393C6.89081 12.6634 6.87463 12.5821 6.87463 12.4999C6.87463 12.4178 6.89081 12.3365 6.92223 12.2606C6.95366 12.1848 6.99972 12.1158 7.05779 12.0578L12.0578 7.05776C12.1158 6.99965 12.1848 6.95355 12.2606 6.9221C12.3365 6.89065 12.4178 6.87446 12.5 6.87446C12.5821 6.87446 12.6634 6.89065 12.7393 6.9221C12.8152 6.95355 12.8841 6.99965 12.9422 7.05776ZM16.8437 3.1562C16.4375 2.7499 15.9552 2.4276 15.4243 2.2077C14.8935 1.98781 14.3246 1.87463 13.75 1.87463C13.1754 1.87463 12.6065 1.98781 12.0756 2.2077C11.5448 2.4276 11.0625 2.7499 10.6562 3.1562L8.30779 5.50385C8.19052 5.62113 8.12463 5.78019 8.12463 5.94604C8.12463 6.1119 8.19052 6.27095 8.30779 6.38823C8.42507 6.50551 8.58413 6.57139 8.74998 6.57139C8.91583 6.57139 9.07489 6.50551 9.19217 6.38823L11.5406 4.04448C12.1288 3.46921 12.9201 3.1491 13.7429 3.15363C14.5656 3.15816 15.3533 3.48697 15.9352 4.06868C16.517 4.6504 16.8459 5.43809 16.8506 6.26082C16.8553 7.08355 16.5353 7.87494 15.9601 8.46323L13.6109 10.8117C13.4936 10.9288 13.4277 11.0878 13.4276 11.2536C13.4276 11.4194 13.4934 11.5784 13.6105 11.6957C13.7277 11.8129 13.8867 11.8789 14.0524 11.8789C14.2182 11.879 14.3772 11.8132 14.4945 11.696L16.8437 9.3437C17.25 8.93744 17.5723 8.45512 17.7922 7.92429C18.0121 7.39346 18.1253 6.82452 18.1253 6.24995C18.1253 5.67538 18.0121 5.10643 17.7922 4.57561C17.5723 4.04478 17.25 3.56246 16.8437 3.1562ZM10.8078 13.6109L8.45935 15.9593C8.17035 16.2548 7.8256 16.49 7.44507 16.6514C7.06454 16.8127 6.65578 16.897 6.24246 16.8992C5.82914 16.9015 5.41948 16.8218 5.03719 16.6646C4.6549 16.5075 4.30758 16.2761 4.01534 15.9838C3.72311 15.6915 3.49176 15.3441 3.33469 14.9618C3.17762 14.5795 3.09795 14.1698 3.1003 13.7565C3.10265 13.3432 3.18697 12.9345 3.34837 12.554C3.50978 12.1734 3.74506 11.8287 4.0406 11.5398L6.38826 9.19214C6.50554 9.07486 6.57142 8.9158 6.57142 8.74995C6.57142 8.5841 6.50554 8.42504 6.38826 8.30776C6.27099 8.19049 6.11193 8.1246 5.94607 8.1246C5.78022 8.1246 5.62116 8.19049 5.50389 8.30776L3.15623 10.6562C2.33572 11.4767 1.87476 12.5896 1.87476 13.7499C1.87476 14.9103 2.33572 16.0232 3.15623 16.8437C3.97674 17.6642 5.0896 18.1252 6.24998 18.1252C7.41036 18.1252 8.52322 17.6642 9.34373 16.8437L11.6922 14.4945C11.8093 14.3772 11.8751 14.2182 11.8751 14.0524C11.875 13.8866 11.8091 13.7277 11.6918 13.6105C11.5745 13.4933 11.4155 13.4275 11.2497 13.4276C11.0839 13.4277 10.925 13.4936 10.8078 13.6109Z"
+                fill="#474D66"
+            />
+        </svg>
+    );
+};
+
+export const LikedIcon = () => {
+    return (
+        <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+                d="M13.9062 3.125C12.293 3.125 10.8805 3.81875 10 4.99141C9.11953 3.81875 7.70703 3.125 6.09375 3.125C4.80955 3.12645 3.57837 3.63723 2.6703 4.5453C1.76223 5.45337 1.25145 6.68455 1.25 7.96875C1.25 13.4375 9.35859 17.8641 9.70391 18.0469C9.79492 18.0958 9.89665 18.1215 10 18.1215C10.1033 18.1215 10.2051 18.0958 10.2961 18.0469C10.6414 17.8641 18.75 13.4375 18.75 7.96875C18.7486 6.68455 18.2378 5.45337 17.3297 4.5453C16.4216 3.63723 15.1904 3.12645 13.9062 3.125ZM10 16.7813C8.57344 15.95 2.5 12.1633 2.5 7.96875C2.50124 7.01601 2.88026 6.10265 3.55396 5.42896C4.22765 4.75526 5.14101 4.37624 6.09375 4.375C7.61328 4.375 8.88906 5.18438 9.42188 6.48438C9.46896 6.59901 9.54907 6.69705 9.65201 6.76605C9.75494 6.83505 9.87607 6.8719 10 6.8719C10.1239 6.8719 10.2451 6.83505 10.348 6.76605C10.4509 6.69705 10.531 6.59901 10.5781 6.48438C11.1109 5.18203 12.3867 4.375 13.9062 4.375C14.859 4.37624 15.7724 4.75526 16.446 5.42896C17.1197 6.10265 17.4988 7.01601 17.5 7.96875C17.5 12.157 11.425 15.9492 10 16.7813Z"
+                fill="#FF6363"
+            />
+        </svg>
+    );
+};
+
+export const ViewIcon = () => {
+    return (
+        <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+                d="M19.3211 9.74688C19.2937 9.68516 18.632 8.21719 17.1609 6.74609C15.2008 4.78594 12.725 3.75 9.99999 3.75C7.27499 3.75 4.79921 4.78594 2.83905 6.74609C1.36796 8.21719 0.703118 9.6875 0.678899 9.74688C0.643362 9.82681 0.625 9.91331 0.625 10.0008C0.625 10.0883 0.643362 10.1748 0.678899 10.2547C0.706243 10.3164 1.36796 11.7836 2.83905 13.2547C4.79921 15.2141 7.27499 16.25 9.99999 16.25C12.725 16.25 15.2008 15.2141 17.1609 13.2547C18.632 11.7836 19.2937 10.3164 19.3211 10.2547C19.3566 10.1748 19.375 10.0883 19.375 10.0008C19.375 9.91331 19.3566 9.82681 19.3211 9.74688ZM9.99999 15C7.5953 15 5.49452 14.1258 3.75546 12.4023C3.0419 11.6927 2.43483 10.8836 1.95312 10C2.4347 9.11636 3.04179 8.30717 3.75546 7.59766C5.49452 5.87422 7.5953 5 9.99999 5C12.4047 5 14.5055 5.87422 16.2445 7.59766C16.9595 8.307 17.5679 9.11619 18.0508 10C17.4875 11.0516 15.0336 15 9.99999 15ZM9.99999 6.25C9.25831 6.25 8.53329 6.46993 7.9166 6.88199C7.29992 7.29404 6.81927 7.87971 6.53544 8.56494C6.25162 9.25016 6.17735 10.0042 6.32205 10.7316C6.46674 11.459 6.82389 12.1272 7.34834 12.6517C7.87279 13.1761 8.54097 13.5333 9.2684 13.6779C9.99583 13.8226 10.7498 13.7484 11.4351 13.4645C12.1203 13.1807 12.7059 12.7001 13.118 12.0834C13.5301 11.4667 13.75 10.7417 13.75 10C13.749 9.00576 13.3535 8.05253 12.6505 7.34949C11.9475 6.64645 10.9942 6.25103 9.99999 6.25ZM9.99999 12.5C9.50554 12.5 9.02219 12.3534 8.61107 12.0787C8.19994 11.804 7.87951 11.4135 7.69029 10.9567C7.50107 10.4999 7.45157 9.99723 7.54803 9.51227C7.64449 9.02732 7.88259 8.58186 8.23222 8.23223C8.58186 7.8826 9.02731 7.6445 9.51227 7.54804C9.99722 7.45157 10.4999 7.50108 10.9567 7.6903C11.4135 7.87952 11.804 8.19995 12.0787 8.61107C12.3534 9.0222 12.5 9.50555 12.5 10C12.5 10.663 12.2366 11.2989 11.7678 11.7678C11.2989 12.2366 10.663 12.5 9.99999 12.5Z"
+                fill="#474D66"
+            />
+        </svg>
+    );
+};
+
+export const SmallUserImgIcon = () => {
+    return (
+        <svg width="28" height="28" viewBox="0 0 28 28" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <rect x="0.5" y="0.5" width="27" height="27" rx="13.5" fill="white" />
+            <rect x="0.5" y="0.5" width="27" height="27" rx="13.5" stroke="#D8DAE5" />
+            <path
+                opacity="0.2"
+                d="M13.125 11.375C13.125 11.7211 13.0224 12.0595 12.8301 12.3472C12.6378 12.635 12.3645 12.8593 12.0447 12.9918C11.7249 13.1242 11.3731 13.1589 11.0336 13.0914C10.6941 13.0239 10.3823 12.8572 10.1376 12.6124C9.89282 12.3677 9.72615 12.0559 9.65863 11.7164C9.5911 11.3769 9.62576 11.0251 9.75821 10.7053C9.89066 10.3855 10.115 10.1122 10.4028 9.91993C10.6905 9.72764 11.0289 9.625 11.375 9.625C11.8391 9.625 12.2842 9.80937 12.6124 10.1376C12.9406 10.4658 13.125 10.9109 13.125 11.375ZM16.625 9.625C16.2789 9.625 15.9405 9.72764 15.6528 9.91993C15.365 10.1122 15.1407 10.3855 15.0082 10.7053C14.8758 11.0251 14.8411 11.3769 14.9086 11.7164C14.9762 12.0559 15.1428 12.3677 15.3876 12.6124C15.6323 12.8572 15.9441 13.0239 16.2836 13.0914C16.6231 13.1589 16.9749 13.1242 17.2947 12.9918C17.6145 12.8593 17.8878 12.635 18.0801 12.3472C18.2724 12.0595 18.375 11.7211 18.375 11.375C18.375 10.9109 18.1906 10.4658 17.8624 10.1376C17.5342 9.80937 17.0891 9.625 16.625 9.625ZM11.375 14.875C11.0289 14.875 10.6905 14.9776 10.4028 15.1699C10.115 15.3622 9.89066 15.6355 9.75821 15.9553C9.62576 16.2751 9.5911 16.6269 9.65863 16.9664C9.72615 17.3059 9.89282 17.6177 10.1376 17.8624C10.3823 18.1072 10.6941 18.2739 11.0336 18.3414C11.3731 18.4089 11.7249 18.3742 12.0447 18.2418C12.3645 18.1093 12.6378 17.885 12.8301 17.5972C13.0224 17.3095 13.125 16.9711 13.125 16.625C13.125 16.1609 12.9406 15.7158 12.6124 15.3876C12.2842 15.0594 11.8391 14.875 11.375 14.875Z"
+                fill="#C1C4D6"
+            />
+            <path
+                d="M11.375 9.1875C10.9424 9.1875 10.5194 9.3158 10.1597 9.55616C9.79996 9.79653 9.51958 10.1382 9.35401 10.5379C9.18845 10.9376 9.14513 11.3774 9.22953 11.8018C9.31394 12.2261 9.52228 12.6159 9.8282 12.9218C10.1341 13.2277 10.5239 13.4361 10.9482 13.5205C11.3726 13.6049 11.8124 13.5616 12.2121 13.396C12.6118 13.2304 12.9535 12.95 13.1938 12.5903C13.4342 12.2306 13.5625 11.8076 13.5625 11.375C13.5625 10.7948 13.332 10.2384 12.9218 9.8282C12.5116 9.41797 11.9552 9.1875 11.375 9.1875ZM11.375 12.6875C11.1154 12.6875 10.8617 12.6105 10.6458 12.4663C10.43 12.3221 10.2617 12.1171 10.1624 11.8773C10.0631 11.6374 10.0371 11.3735 10.0877 11.1189C10.1384 10.8643 10.2634 10.6305 10.4469 10.4469C10.6305 10.2634 10.8643 10.1384 11.1189 10.0877C11.3735 10.0371 11.6374 10.0631 11.8773 10.1624C12.1171 10.2617 12.3221 10.43 12.4663 10.6458C12.6105 10.8617 12.6875 11.1154 12.6875 11.375C12.6875 11.7231 12.5492 12.0569 12.3031 12.3031C12.0569 12.5492 11.7231 12.6875 11.375 12.6875ZM16.625 13.5625C17.0576 13.5625 17.4806 13.4342 17.8403 13.1938C18.2 12.9535 18.4804 12.6118 18.646 12.2121C18.8116 11.8124 18.8549 11.3726 18.7705 10.9482C18.6861 10.5239 18.4777 10.1341 18.1718 9.8282C17.8659 9.52228 17.4761 9.31394 17.0518 9.22953C16.6274 9.14513 16.1876 9.18845 15.7879 9.35401C15.3882 9.51958 15.0465 9.79996 14.8062 10.1597C14.5658 10.5194 14.4375 10.9424 14.4375 11.375C14.4375 11.9552 14.668 12.5116 15.0782 12.9218C15.4884 13.332 16.0448 13.5625 16.625 13.5625ZM16.625 10.0625C16.8846 10.0625 17.1383 10.1395 17.3542 10.2837C17.57 10.4279 17.7383 10.6329 17.8376 10.8727C17.9369 11.1126 17.9629 11.3765 17.9123 11.6311C17.8616 11.8857 17.7366 12.1195 17.5531 12.3031C17.3695 12.4866 17.1357 12.6116 16.8811 12.6623C16.6265 12.7129 16.3626 12.6869 16.1227 12.5876C15.8829 12.4883 15.6779 12.32 15.5337 12.1042C15.3895 11.8883 15.3125 11.6346 15.3125 11.375C15.3125 11.0269 15.4508 10.6931 15.6969 10.4469C15.9431 10.2008 16.2769 10.0625 16.625 10.0625ZM11.375 14.4375C10.9424 14.4375 10.5194 14.5658 10.1597 14.8062C9.79996 15.0465 9.51958 15.3882 9.35401 15.7879C9.18845 16.1876 9.14513 16.6274 9.22953 17.0518C9.31394 17.4761 9.52228 17.8659 9.8282 18.1718C10.1341 18.4777 10.5239 18.6861 10.9482 18.7705C11.3726 18.8549 11.8124 18.8116 12.2121 18.646C12.6118 18.4804 12.9535 18.2 13.1938 17.8403C13.4342 17.4806 13.5625 17.0576 13.5625 16.625C13.5625 16.0448 13.332 15.4884 12.9218 15.0782C12.5116 14.668 11.9552 14.4375 11.375 14.4375ZM11.375 17.9375C11.1154 17.9375 10.8617 17.8605 10.6458 17.7163C10.43 17.5721 10.2617 17.3671 10.1624 17.1273C10.0631 16.8874 10.0371 16.6235 10.0877 16.3689C10.1384 16.1143 10.2634 15.8805 10.4469 15.6969C10.6305 15.5134 10.8643 15.3884 11.1189 15.3377C11.3735 15.2871 11.6374 15.3131 11.8773 15.4124C12.1171 15.5117 12.3221 15.68 12.4663 15.8958C12.6105 16.1117 12.6875 16.3654 12.6875 16.625C12.6875 16.9731 12.5492 17.3069 12.3031 17.5531C12.0569 17.7992 11.7231 17.9375 11.375 17.9375ZM18.8125 16.625C18.8125 16.741 18.7664 16.8523 18.6844 16.9344C18.6023 17.0164 18.491 17.0625 18.375 17.0625H17.0625V18.375C17.0625 18.491 17.0164 18.6023 16.9344 18.6844C16.8523 18.7664 16.741 18.8125 16.625 18.8125C16.509 18.8125 16.3977 18.7664 16.3156 18.6844C16.2336 18.6023 16.1875 18.491 16.1875 18.375V17.0625H14.875C14.759 17.0625 14.6477 17.0164 14.5656 16.9344C14.4836 16.8523 14.4375 16.741 14.4375 16.625C14.4375 16.509 14.4836 16.3977 14.5656 16.3156C14.6477 16.2336 14.759 16.1875 14.875 16.1875H16.1875V14.875C16.1875 14.759 16.2336 14.6477 16.3156 14.5656C16.3977 14.4836 16.509 14.4375 16.625 14.4375C16.741 14.4375 16.8523 14.4836 16.9344 14.5656C17.0164 14.6477 17.0625 14.759 17.0625 14.875V16.1875H18.375C18.491 16.1875 18.6023 16.2336 18.6844 16.3156C18.7664 16.3977 18.8125 16.509 18.8125 16.625Z"
+                fill="#C1C4D6"
+            />
+        </svg>
+    );
+};
+
+export const UserImgIcon = () => {
+    return (
+        <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <rect x="0.714286" y="0.714286" width="38.5714" height="38.5714" rx="19.2857" fill="white" />
+            <rect x="0.714286" y="0.714286" width="38.5714" height="38.5714" rx="19.2857" stroke="#D8DAE5" strokeWidth="1.42857" />
+            <path
+                opacity="0.2"
+                d="M18.75 16.25C18.75 16.7445 18.6034 17.2278 18.3287 17.6389C18.054 18.05 17.6635 18.3705 17.2067 18.5597C16.7499 18.7489 16.2472 18.7984 15.7623 18.702C15.2773 18.6055 14.8319 18.3674 14.4822 18.0178C14.1326 17.6681 13.8945 17.2227 13.798 16.7377C13.7016 16.2528 13.7511 15.7501 13.9403 15.2933C14.1295 14.8365 14.45 14.446 14.8611 14.1713C15.2722 13.8966 15.7555 13.75 16.25 13.75C16.913 13.75 17.5489 14.0134 18.0178 14.4822C18.4866 14.9511 18.75 15.587 18.75 16.25ZM23.75 13.75C23.2555 13.75 22.7722 13.8966 22.3611 14.1713C21.95 14.446 21.6295 14.8365 21.4403 15.2933C21.2511 15.7501 21.2016 16.2528 21.298 16.7377C21.3945 17.2227 21.6326 17.6681 21.9822 18.0178C22.3319 18.3674 22.7773 18.6055 23.2623 18.702C23.7472 18.7984 24.2499 18.7489 24.7067 18.5597C25.1635 18.3705 25.554 18.05 25.8287 17.6389C26.1034 17.2278 26.25 16.7445 26.25 16.25C26.25 15.587 25.9866 14.9511 25.5178 14.4822C25.0489 14.0134 24.413 13.75 23.75 13.75ZM16.25 21.25C15.7555 21.25 15.2722 21.3966 14.8611 21.6713C14.45 21.946 14.1295 22.3365 13.9403 22.7933C13.7511 23.2501 13.7016 23.7528 13.798 24.2377C13.8945 24.7227 14.1326 25.1681 14.4822 25.5178C14.8319 25.8674 15.2773 26.1055 15.7623 26.202C16.2472 26.2984 16.7499 26.2489 17.2067 26.0597C17.6635 25.8705 18.054 25.55 18.3287 25.1389C18.6034 24.7278 18.75 24.2445 18.75 23.75C18.75 23.087 18.4866 22.4511 18.0178 21.9822C17.5489 21.5134 16.913 21.25 16.25 21.25Z"
+                fill="#C1C4D6"
+            />
+            <path
+                d="M16.25 13.125C15.6319 13.125 15.0277 13.3083 14.5138 13.6517C13.9999 13.995 13.5994 14.4831 13.3629 15.0541C13.1264 15.6251 13.0645 16.2535 13.185 16.8597C13.3056 17.4658 13.6033 18.0227 14.0403 18.4597C14.4773 18.8967 15.0342 19.1944 15.6403 19.315C16.2465 19.4355 16.8749 19.3736 17.4459 19.1371C18.0169 18.9006 18.505 18.5001 18.8483 17.9862C19.1917 17.4723 19.375 16.8681 19.375 16.25C19.375 15.4212 19.0458 14.6263 18.4597 14.0403C17.8737 13.4542 17.0788 13.125 16.25 13.125ZM16.25 18.125C15.8792 18.125 15.5166 18.015 15.2083 17.809C14.9 17.603 14.6596 17.3101 14.5177 16.9675C14.3758 16.6249 14.3387 16.2479 14.411 15.8842C14.4834 15.5205 14.662 15.1864 14.9242 14.9242C15.1864 14.662 15.5205 14.4834 15.8842 14.411C16.2479 14.3387 16.6249 14.3758 16.9675 14.5177C17.3101 14.6596 17.603 14.9 17.809 15.2083C18.015 15.5166 18.125 15.8792 18.125 16.25C18.125 16.7473 17.9275 17.2242 17.5758 17.5758C17.2242 17.9275 16.7473 18.125 16.25 18.125ZM23.75 19.375C24.3681 19.375 24.9723 19.1917 25.4862 18.8483C26.0001 18.505 26.4006 18.0169 26.6371 17.4459C26.8737 16.8749 26.9355 16.2465 26.815 15.6403C26.6944 15.0342 26.3968 14.4773 25.9597 14.0403C25.5227 13.6033 24.9659 13.3056 24.3597 13.185C23.7535 13.0645 23.1251 13.1264 22.5541 13.3629C21.9831 13.5994 21.495 13.9999 21.1517 14.5138C20.8083 15.0277 20.625 15.6319 20.625 16.25C20.625 17.0788 20.9542 17.8737 21.5403 18.4597C22.1263 19.0458 22.9212 19.375 23.75 19.375ZM23.75 14.375C24.1208 14.375 24.4834 14.485 24.7917 14.691C25.1 14.897 25.3404 15.1899 25.4823 15.5325C25.6242 15.8751 25.6613 16.2521 25.589 16.6158C25.5166 16.9795 25.3381 17.3136 25.0758 17.5758C24.8136 17.8381 24.4795 18.0166 24.1158 18.089C23.7521 18.1613 23.3751 18.1242 23.0325 17.9823C22.6899 17.8404 22.397 17.6 22.191 17.2917C21.985 16.9834 21.875 16.6208 21.875 16.25C21.875 15.7527 22.0725 15.2758 22.4242 14.9242C22.7758 14.5725 23.2527 14.375 23.75 14.375ZM16.25 20.625C15.6319 20.625 15.0277 20.8083 14.5138 21.1517C13.9999 21.495 13.5994 21.9831 13.3629 22.5541C13.1264 23.1251 13.0645 23.7535 13.185 24.3597C13.3056 24.9659 13.6033 25.5227 14.0403 25.9597C14.4773 26.3968 15.0342 26.6944 15.6403 26.815C16.2465 26.9355 16.8749 26.8737 17.4459 26.6371C18.0169 26.4006 18.505 26.0001 18.8483 25.4862C19.1917 24.9723 19.375 24.3681 19.375 23.75C19.375 22.9212 19.0458 22.1263 18.4597 21.5403C17.8737 20.9542 17.0788 20.625 16.25 20.625ZM16.25 25.625C15.8792 25.625 15.5166 25.515 15.2083 25.309C14.9 25.103 14.6596 24.8101 14.5177 24.4675C14.3758 24.1249 14.3387 23.7479 14.411 23.3842C14.4834 23.0205 14.662 22.6864 14.9242 22.4242C15.1864 22.162 15.5205 21.9834 15.8842 21.911C16.2479 21.8387 16.6249 21.8758 16.9675 22.0177C17.3101 22.1596 17.603 22.4 17.809 22.7083C18.015 23.0167 18.125 23.3792 18.125 23.75C18.125 24.2473 17.9275 24.7242 17.5758 25.0758C17.2242 25.4275 16.7473 25.625 16.25 25.625ZM26.875 23.75C26.875 23.9158 26.8092 24.0747 26.6919 24.1919C26.5747 24.3092 26.4158 24.375 26.25 24.375H24.375V26.25C24.375 26.4158 24.3092 26.5747 24.1919 26.6919C24.0747 26.8092 23.9158 26.875 23.75 26.875C23.5842 26.875 23.4253 26.8092 23.3081 26.6919C23.1908 26.5747 23.125 26.4158 23.125 26.25V24.375H21.25C21.0842 24.375 20.9253 24.3092 20.8081 24.1919C20.6908 24.0747 20.625 23.9158 20.625 23.75C20.625 23.5842 20.6908 23.4253 20.8081 23.3081C20.9253 23.1908 21.0842 23.125 21.25 23.125H23.125V21.25C23.125 21.0842 23.1908 20.9253 23.3081 20.8081C23.4253 20.6908 23.5842 20.625 23.75 20.625C23.9158 20.625 24.0747 20.6908 24.1919 20.8081C24.3092 20.9253 24.375 21.0842 24.375 21.25V23.125H26.25C26.4158 23.125 26.5747 23.1908 26.6919 23.3081C26.8092 23.4253 26.875 23.5842 26.875 23.75Z"
+                fill="#C1C4D6"
+            />
+        </svg>
+    );
+};
+
+export const HamburgerIcon = () => {
+    return (
+        <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path
+                d="M13.125 12C13.125 12.2225 13.059 12.44 12.9354 12.625C12.8118 12.81 12.6361 12.9542 12.4305 13.0394C12.225 13.1245 11.9988 13.1468 11.7805 13.1034C11.5623 13.06 11.3618 12.9528 11.2045 12.7955C11.0472 12.6382 10.94 12.4377 10.8966 12.2195C10.8532 12.0012 10.8755 11.775 10.9606 11.5695C11.0458 11.3639 11.19 11.1882 11.375 11.0646C11.56 10.941 11.7775 10.875 12 10.875C12.2984 10.875 12.5845 10.9935 12.7955 11.2045C13.0065 11.4155 13.125 11.7016 13.125 12ZM12 6.75C12.2225 6.75 12.44 6.68402 12.625 6.5604C12.81 6.43679 12.9542 6.26109 13.0394 6.05552C13.1245 5.84995 13.1468 5.62375 13.1034 5.40552C13.06 5.1873 12.9528 4.98684 12.7955 4.82951C12.6382 4.67217 12.4377 4.56503 12.2195 4.52162C12.0012 4.47821 11.775 4.50049 11.5695 4.58564C11.3639 4.67078 11.1882 4.81498 11.0646 4.99998C10.941 5.18499 10.875 5.4025 10.875 5.625C10.875 5.92337 10.9935 6.20952 11.2045 6.4205C11.4155 6.63147 11.7016 6.75 12 6.75ZM12 17.25C11.7775 17.25 11.56 17.316 11.375 17.4396C11.19 17.5632 11.0458 17.7389 10.9606 17.9445C10.8755 18.15 10.8532 18.3762 10.8966 18.5945C10.94 18.8127 11.0472 19.0132 11.2045 19.1705C11.3618 19.3278 11.5623 19.435 11.7805 19.4784C11.9988 19.5218 12.225 19.4995 12.4305 19.4144C12.6361 19.3292 12.8118 19.185 12.9354 19C13.059 18.815 13.125 18.5975 13.125 18.375C13.125 18.0766 13.0065 17.7905 12.7955 17.5795C12.5845 17.3685 12.2984 17.25 12 17.25Z"
+                fill="#474D66"
+            />
+        </svg>
+    );
+};
+
+export const BottomArrowIcon = ({ isOpen }: { isOpen: boolean }) => {
+    return (
+        <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg" style={{ transform: `rotate(${isOpen ? 0 : -180}deg)`, transition: "transform 0.2s" }}>
+            <path
+                d="M13.3538 6.35378L8.35378 11.3538C8.30735 11.4003 8.2522 11.4372 8.1915 11.4623C8.13081 11.4875 8.06574 11.5004 8.00003 11.5004C7.93433 11.5004 7.86926 11.4875 7.80856 11.4623C7.74786 11.4372 7.69272 11.4003 7.64628 11.3538L2.64628 6.35378C2.55246 6.25996 2.49976 6.13272 2.49976 6.00003C2.49976 5.86735 2.55246 5.7401 2.64628 5.64628C2.7401 5.55246 2.86735 5.49976 3.00003 5.49976C3.13272 5.49976 3.25996 5.55246 3.35378 5.64628L8.00003 10.2932L12.6463 5.64628C12.6927 5.59983 12.7479 5.56298 12.8086 5.53784C12.8693 5.5127 12.9343 5.49976 13 5.49976C13.0657 5.49976 13.1308 5.5127 13.1915 5.53784C13.2522 5.56298 13.3073 5.59983 13.3538 5.64628C13.4002 5.69274 13.4371 5.74789 13.4622 5.80859C13.4874 5.86928 13.5003 5.93434 13.5003 6.00003C13.5003 6.06573 13.4874 6.13079 13.4622 6.19148C13.4371 6.25218 13.4002 6.30733 13.3538 6.35378Z"
+                fill="#8F95B2"
+            />
+        </svg>
+    );
+};
+
+export const LockIcon = () => {
+    return (
+        <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg" className="cursor-pointer">
+            <path
+                d="M10 1.875C8.39303 1.875 6.82214 2.35152 5.486 3.24431C4.14985 4.1371 3.10844 5.40605 2.49348 6.8907C1.87852 8.37535 1.71762 10.009 2.03112 11.5851C2.34463 13.1612 3.11846 14.6089 4.25476 15.7452C5.39106 16.8815 6.8388 17.6554 8.4149 17.9689C9.99099 18.2824 11.6247 18.1215 13.1093 17.5065C14.594 16.8916 15.8629 15.8502 16.7557 14.514C17.6485 13.1779 18.125 11.607 18.125 10C18.1227 7.84581 17.266 5.78051 15.7427 4.25727C14.2195 2.73403 12.1542 1.87727 10 1.875ZM10 16.875C8.64026 16.875 7.31105 16.4718 6.18046 15.7164C5.04987 14.9609 4.16868 13.8872 3.64833 12.6309C3.12798 11.3747 2.99183 9.99237 3.2571 8.65875C3.52238 7.32513 4.17716 6.10013 5.13864 5.13864C6.10013 4.17716 7.32514 3.52237 8.65876 3.2571C9.99238 2.99183 11.3747 3.12798 12.631 3.64833C13.8872 4.16868 14.9609 5.04987 15.7164 6.18045C16.4718 7.31104 16.875 8.64025 16.875 10C16.8729 11.8227 16.1479 13.5702 14.8591 14.8591C13.5702 16.1479 11.8227 16.8729 10 16.875ZM9.375 10.625V6.25C9.375 6.08424 9.44085 5.92527 9.55806 5.80806C9.67527 5.69085 9.83424 5.625 10 5.625C10.1658 5.625 10.3247 5.69085 10.4419 5.80806C10.5592 5.92527 10.625 6.08424 10.625 6.25V10.625C10.625 10.7908 10.5592 10.9497 10.4419 11.0669C10.3247 11.1842 10.1658 11.25 10 11.25C9.83424 11.25 9.67527 11.1842 9.55806 11.0669C9.44085 10.9497 9.375 10.7908 9.375 10.625ZM10.9375 13.4375C10.9375 13.6229 10.8825 13.8042 10.7795 13.9583C10.6765 14.1125 10.5301 14.2327 10.3588 14.3036C10.1875 14.3746 9.99896 14.3932 9.81711 14.357C9.63525 14.3208 9.4682 14.2315 9.33709 14.1004C9.20598 13.9693 9.11669 13.8023 9.08052 13.6204C9.04434 13.4385 9.06291 13.25 9.13387 13.0787C9.20482 12.9074 9.32499 12.761 9.47916 12.658C9.63333 12.555 9.81458 12.5 10 12.5C10.2486 12.5 10.4871 12.5988 10.6629 12.7746C10.8387 12.9504 10.9375 13.1889 10.9375 13.4375Z"
+                fill="#8F95B2"
+            />
+        </svg>
+    );
+};
